+++ Task2/src/school/Subject.java
@@ -4,7 +4,7 @@ public enum Subject {
 
 	MATHS, PHYSICS, LITERATURE, HISTORY, CHEMISTRY, ART, MUSIC;
 	
-	public static Subject showSubject(String str) {
+	public static Subject getSubject(String str) {
 		Subject[] values = Subject.values();
 		
 		for (Subject value : values) {
